Stop the claimer tests racing the poller, and get the CLI token exchange off a held connection
**The tests.** The claimer e2e inserted rows due one second ago and then raced
to claim them -- against a session-scoped worker whose poller claims anything
due at real now, continuously, from the same database. They passed only because
`--dist loadscope` happened to run that file before anything that starts a
worker. Test order as a load-bearing assumption is the exact shape this branch
keeps finding and removing, so it should not survive in the tests written to
prove the branch works.

The claiming tests now run six hours ahead: rows are due at now+6h and each test
passes its own `now` six hours forward. The real poller sees nothing due; the
explicit `now=` sees everything. Verified by running the claimer suites *after*
a test that starts a worker -- the exact ordering that was racing -- 16 passed.

The backfill tests deliberately stay on real time, with a note saying why:
backfilling is convergent, so it does not matter who got there first, and their
assertions hold either way. That is a stronger property than "no poller was
running". One assertion did have to be scoped -- "no schedule.fired rows exist"
was testing the whole database and would have failed on someone else's
legitimate fire.

**The CLI routes.** `/auth/cli/session-tokens` and `/auth/cli/refresh` both took
a request-scoped `UserServiceDep` and then made a SuperTokens round trip with
that connection checked out. On `/cli/refresh` the exchange happens *before* the
database is consulted at all, so the connection was held across it for nothing.
Each now opens a short scope around its one lookup.

That pushed `auth_controller.py` from 595 to 609 lines, over the ratchet's 600
ceiling -- so the two routes moved to `cli_auth_controller.py` with their
request/response models, which is where they belonged anyway: a CLI-only
surface, deliberately absent from the public schema. Extraction is the way past
the ceiling; re-baselining is not. The file lands at 522.

Neither route has any test coverage, which is worth saying out loud. Verified by
asserting both are still registered on the module's routers and that FastAPI can
build their dependants.
